Updated fields which require monetary values from INT to NUMERIC(12,2)

+++ server/database.sql
@@ -30,7 +30,7 @@ CREATE TABLE auction (
     seller_id INT NOT NULL, --profile of the seller
     name VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
-    opening_bid INT NOT NULL,
+    opening_bid NUMERIC(12,2) NOT NULL,
     closing_date TIMESTAMP NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (seller_id) REFERENCES profile (id) ON DELETE CASCADE ON UPDATE CASCADE --auctions tied to profile not the user
@@ -40,7 +40,7 @@ CREATE TABLE bid (
     id SERIAL PRIMARY KEY,
     bidder_id INT NOT NULL,
     auction_id INT NOT NULL,
-    amount INT NOT NULL,
+    amount NUMERIC(12,2) NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (bidder_id) REFERENCES app_user (id) ON DELETE CASCADE ON UPDATE CASCADE,
     FOREIGN KEY (auction_id) REFERENCES auction (id) ON DELETE CASCADE ON UPDATE CASCADE
